# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервис, который является частью системы геооповещений",
        "title": "geonotify-service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/v1/incidents/stats": {
            "get": {
                "description": "Получить статистику уникальных пользователей за последние N минут",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Статистика по зонам",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_resp.StatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/location/check": {
            "post": {
                "description": "Проверить, попадает ли точка в опасную зону (публичный эндпоинт)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Проверить координаты",
                "parameters": [
                    {
                        "description": "Координаты для проверки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_req.LocationCheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_resp.LocationCheckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_4otis_geonotify-service_internal_dto_req.LocationCheckRequest": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_4otis_geonotify-service_internal_dto_resp.IncidentResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "incident_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "radius_m": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_4otis_geonotify-service_internal_dto_resp.LocationCheckResponse": {
            "type": "object",
            "properties": {
                "has_alert": {
                    "type": "boolean"
                },
                "incidents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_resp.IncidentResponse"
                    }
                }
            }
        },
        "github_com_4otis_geonotify-service_internal_dto_resp.StatsResponse": {
            "type": "object",
            "properties": {
                "period_start": {
                    "type": "string"
                },
                "total_checks": {
                    "type": "integer"
                },
                "user_count": {
                    "type": "integer"
                },
                "window_minutes": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}